Dockerfile: removed device-tree-compiler download and install
after further review, device-tree-compiler is only used for sanity
check on own dtc implementation, and uses local if version 1.4.7 or
above.

But, if not installed, uses the one in zephyr SDK (1.5.0-dirty), so
this one is redundant.

The absolute path download is volatile and often fails, removing it
will solve future build breaks.

Signed-off-by: John Grey <greyjohn734@gmail.com>