Dockerfile: Do not pull sources in advance
This reverts a change from #38 which moved all wget commands to the start of the Dockerfile.
The change had the advantage that build errors caused by missing sources occur earlier.
However, this causes the sources to be added and deleted again in different layers.
Thus, the sources remain part of the Docker image.
This leads to a size increase of the compressed image from 2GB to 3.6 GB.

Signed-off-by: Maximilian Bachmann <m.bachmann@acontis.com>