Remove `ZEPHYR_SDK_INSTALL_DIR` environment variable
This commit removes the `ZEPHYR_SDK_INSTALL_DIR` environment variable,
which has the effect of forcing a specific version of Zephyr SDK to be
used.

When `ZEPHYR_SDK_INSTALL_DIR` is not set, the Zephyr build system will
automatically detect the required version of Zephyr SDK (installed in
the Docker image) and use it.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>