Install Zephyr SDK CMake package
This commit adds the `-y` option when invoking the Zephyr SDK installer
so that the Zephyr SDK CMake package is installed.

This ensures that the Zephyr build system can locate the required
version of Zephyr SDK even when it is not installed in one of the
default directories specified as PATHS of the `find_package` command.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>