Dockerfile: Add Doxygen installation step
This commit adds a step that installs a specific version of Doxygen
binaries because the distro-provided Doxygen package may be outdated
and not the version required for building Zephyr documentation.

Note that the custom Doxygen binaries are only installed for the x86-64
build of the Docker image because the official Doxygen website only
provides the x86-64 binaries.

For this reason, the distro-provided Doxygen package, which is
available for multiple architectures, is still installed so that
Doxygen is available for non-x86 images -- in case of the x86-64 image,
which contains both the distro Doxygen package and the custom Doxygen
installation, the latter takes precedence because it is sym-linked in
`/usr/local/bin`, which takes precedence over `/usr/bin` in PATH.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>