Dockerfile: Remove awscli Python package
This commit removes the `awscli` Python package from the base CI image
because it is not currently not used by any Zephyr CI workflows and has
`PyYAML` dependency conflict with the Renode (see the GH issue #145).

If the AWS command line tool is necessary in the future, look into
either migrating to the `awscliv2` or installing the `awscli` package
in a virtual environment.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>