.github: workflows: Disable ARM64 image build
This commit temporarily disables the ARM64 Docker image build because
the Zephyr AArch64 CI runners are currently unavailable and the current
Zephyr SDK 0.16.5 release does not include an ARM64 build.

Revert this commit when the AArch64 CI runners are fixed and a new
Zephyr SDK with ARM64 build is released.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>