ci: Add Kitware ninja v1.11.1
This commit adds the Kitware ninja v1.11.1 to the CI Docker image.

The Kitware fork of ninja is required to enable the "jobserver" feature,
which is currently not available in the upstream ninja codebase.

The ninja jobserver client feature is used by the Zephyr CI workflow in
conjunction with the twister-side jobserver to globally limit the number
of parallel compilations, which is necessary to keep the runner RAM
usage under control.

Note that Kitware ninja is currently only included in the x86-64 image
because pre-compiled AArch64 binary is not available -- for AArch64, the
distro-provided vanilla ninja will be used.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>