.github: workflows: Use zephyr-runner v2
This commit updates the CI workflow to use the new zephyr-runner v2.

As part of this, it also reworks the workflow to use Buildah and Podman,
which is daemon-less and does not require root privileges, to build
Docker images because the new runners have more strict security
requirements and run workflow pods in unprivileged mode, which
effectively makes it impossible to run Docker daemon inside a runner.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>