ci: Place container storage inside work volume
This commit updates the CI workflow to explicitly configure the Podman
container storage to be inside the runner work volume (`__w`), which is
guaranteed to have at least 100GB available.

The default container storage path is `/var/lib/containers/storage`,
which resides in the Docker storage volume that has no free space
guarantees.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>