ci: Add rust target support
Invoke `rust target add ...` for each of the targets supported by
Zephyr.  There is a bit of a catch-22 here.  We should be getting this
list from the build, but the change that adds Rust support can't merge
until Rust support has been tested.

Once the first PR is merged, and the zephyr/lib/rust/targets.txt file
exists, the hardcoded list here can be replaced by a loop that adds each
of these targest.

All of the targets adds about 300Mb to the image, as these targets only
include a fairly minimal 'core' and 'alloc' library.

Signed-off-by: David Brown <david.brown@linaro.org>