base: Install libc6-dbg:i386
In tests, users may see a segfault in CI which they may not be able to
reproduce locally (due to having a different OS).
A user may try to get more info in the CI by running the test with
valgrind. But for valgrind to run with 32bit targets we need
libc6-dbg:i386 installed.
So let's install it by default so it is ready for users.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>