fix: pin pygobject pip package to 3.50.0
Versions of pygobject from 3.51.0 forward require
libgirepository-2.0-dev as a dependency, but that one is not yet
available on the Ubuntu 22.04 (it is on 24.04).

3.50.0 version requires libgirepository-1.0-dev, which is available on
the Ubuntu 22.04, however that package was never explicitly installed,
so this commit also adds it.

By pinning the pip package and adding new apt package we ensure that
Dockerfile.base can be even built.

Closes: #222