base: Fix i386 APT package dependency hell
One of the i386 packages installed in the base Docker image depends on
`python3-minimal` and APT tries to nuke the previously installed `python3`
and all its friends, including `software-properties-common`, during the
i386 package installation.

Marking `python3` and `python3-apt` as held seems to make APT not want to
do that, and keeps `python3` and its friends in place without any hacks.

This should, hopefully, be the ultimate fix for all the i386 package
dependency issues we have come across in the past.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>