Dockerfile: Move all components except Zephyr SDK to the CI base image
Until now, the logical separation between the `ci-base` and `ci` images was
not so well defined and the components were not organised into each image
in a practically useful way.

After this change, the `ci-base` image is defined to be an image that
includes all tools required for CI operation except the Zephyr SDK, and the
`ci` image is defined as an image that includes everything in the `ci-base`
image plus the Zephyr SDK.

This effectively allows the CI workflows that do not require the Zephyr SDK
(e.g. CI workflows for third-party toolchains, or SDK CI workflow that
tests an arbitrary SDK build) to use the `ci-base` image.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>